+++ hw12_13_14_15_calendar/cmd/calendar/main.go
@@ -6,7 +6,7 @@ import (
 	"flag"
 	"fmt"
 
-	"github.com/dmitryt/otus-golang-hw/hw12_13_14_15_calendar/internal/calendar"
+	"github.com/dmitryt/otus-golang-hw/hw12_13_14_15_calendar/internal/app"
 	"github.com/dmitryt/otus-golang-hw/hw12_13_14_15_calendar/internal/config"
 	"github.com/dmitryt/otus-golang-hw/hw12_13_14_15_calendar/internal/logger"
 	"github.com/dmitryt/otus-golang-hw/hw12_13_14_15_calendar/internal/repository"
@@ -49,7 +49,7 @@ func main() {
 	}
 	defer repo.Close()
 
-	app, err := calendar.New(repo)
+	app, err := app.New(repo)
 	if err != nil {
 		fatal(err)
 	}
