+++ hw12_13_14_15_calendar/internal/app/errors.go
@@ -1,4 +1,4 @@
-package calendar
+package app
 
 import (
 	"errors"
